factor out the SonosController initialization into itself, so we can have a refresh method on the controller.
This takes the work for discovering SonosController metadata, and moves it into the SonosController class. This allows us to have a refresh method. That way at any time, we can call refresh, to have the controller refresh it's information, like its group, without having to do a full discover.

With the refresh method being in place, the portion that fetches metadata for each found device, was replaced with simply creating the controller with the IP address and port found after discovery, and calling refresh on it.